ml_ensemble_microservices/requirements.txt
streamlit>=1.31.0
databricks-sdk>=0.18.0
pyyaml>=6.0
mlflow>=2.9.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
